Added page for Bedroom Data. Bedroom page added to route. Icon has been added to Navigation Footer. User has been taken out of Navigation Footer and will be eventually moved to the Home screen or elsewhere.
