+++ src/components/NavigationFooter/NavigationFooter.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import * as styles from "./NavigationFooter.module.scss";
 import { Link, useLocation } from "react-router-dom";
 import { AiFillHome } from "react-icons/ai";
-import { FaUser } from "react-icons/fa";
+import { FaUser, FaThermometerThreeQuarters } from "react-icons/fa";
 import { RiHealthBookFill } from "react-icons/ri";
 import { GiNightSleep } from "react-icons/gi";
 import { GoFlame } from "react-icons/go";
@@ -15,12 +15,12 @@ const NavigationHeader = () => {
       <Link className={styles.link} to="/">
         <AiFillHome size={35} color={pathname === "/" ? "#66becc" : null} />
       </Link>
-      <Link className={styles.link} to="/user-data">
+      {/* <Link className={styles.link} to="/user-data">
         <FaUser
           size={35}
           color={pathname === "/user-data" ? "#66becc" : null}
         />
-      </Link>
+      </Link> */}
       <Link className={styles.link} to="/readiness-data">
         <RiHealthBookFill
           size={35}
@@ -37,6 +37,12 @@ const NavigationHeader = () => {
         <GoFlame
           size={35}
           color={pathname === "/activity-data" ? "#66becc" : null}
+        />
+      </Link>
+      <Link className={styles.link} to="/bedroom-data">
+        <FaThermometerThreeQuarters
+          size={35}
+          color={pathname === "/bedroom-data" ? "#66becc" : null}
         />
       </Link>
     </div>
