+++ src/components/RenderBedroomData/RenderBedroomData.module.scss
@@ -1,0 +1,5 @@
+.renderBedroomContainer {
+  text-align: center;
+  font-size: 1.6rem;
+  font-weight: 200;
+}
